Chatroom CSS
Minor HTML edits to the chatroom.html, and the addition of the
chatroom.css. Pardon the white background, but I wasn's sure I wanted it
set to another solid color. Known issue: the text bar at the bottom of
the page will cover the last message on the page if enough messages are
in the debate. This is fixable by inserting a few <br> for now, but
needs a better fix later.